refactor: type 선언을 분리한다
