+++ src/entities/payment/api.ts
@@ -1,42 +1,29 @@
-import { type HttpClient } from '../../shared/api/http-client';
-
-export type Order = {
-  id: string;
-  amount: number;
-  orderName: string;
-};
-
-export type RedirectURL = {
-  successUrl: string;
-  failUrl: string;
-};
-
-export type PaymentApi = {
-  requestPayment: (
-    body: Order & RedirectURL,
-  ) => Promise<{ token: string; expiredAt: string }>;
-  subscribePaymentEvents: (token: string) => EventSource;
-};
+import type { HttpClient } from '../../shared/api/httpClient';
+import { API_ENDPOINTS } from '../../shared/config/apiEndpoints';
+import type { PaymentApi, PaymentReq, PaymentRes } from './types';
 
 export const createPaymentApi = (httpClient: HttpClient): PaymentApi => {
   return {
     requestPayment: async (body) => {
-      const response = await httpClient.post<
-        Order & RedirectURL,
-        { token: string; expiredAt: string }
-      >('/api/v1/payments', body);
+      const response = await httpClient.post<PaymentReq, PaymentRes>(
+        API_ENDPOINTS.REQUEST_PAYMENT,
+        body,
+      );
 
       if (response.code === 200) {
-        return response.data;
+        return response;
       } else {
-        throw new Error(response.message || 'Payment request failed');
+        throw new Error(response.error?.message || 'Payment request failed');
       }
     },
 
     subscribePaymentEvents: (token: string): EventSource => {
-      const eventSource = new EventSource(`/api/v1/payments/${token}/events`, {
-        withCredentials: true,
-      });
+      const eventSource = new EventSource(
+        API_ENDPOINTS.SUBSCRIBE_PAYMENT_EVENTS(token),
+        {
+          withCredentials: true,
+        },
+      );
 
       return eventSource;
     },
